% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_allnodes_alledges.R
\name{make_allnodes_alledges}
\alias{make_allnodes_alledges}
\title{Make Dataframe with All Nodes and All Edges}
\usage{
make_allnodes_alledges(pw_biopax, exclude_ids = NULL, verbose = FALSE)
}
\arguments{
\item{pw_biopax}{Pathway as a BioPAX object.}

\item{exclude_ids}{Component ids to exclude from graph.}

\item{verbose}{Boolean. Show all warnings?}
}
\description{
Make fully annotated dataframes with all nodes and all edges required to make a graph.
}
\author{
Ivan Grishagin
}
