% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_RIGplotbiopax.R
\name{update_RIGplotbiopax}
\alias{update_RIGplotbiopax}
\title{Update RIGplotbiopax from GitHub}
\usage{
update_RIGplotbiopax()
}
\description{
Updates the package. It may be necessary to restart the R session after the update.
}
\author{
Ivan Grishagin
}
