% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_allnodes_alledges.R
\name{make_allnodes_alledges}
\alias{make_allnodes_alledges}
\title{Make Dataframe with All Nodes and All Edges}
\usage{
make_allnodes_alledges(nodes_list, edges_list, exclude_ids = NULL)
}
\arguments{
\item{nodes_list}{List containing dataframes with all nodes.}

\item{edges_list}{List containing dataframes with all edges.}

\item{exclude_ids}{Component ids to exclude from graph.}
}
\description{
Make fully annotated dataframes with all nodes and all edges required to make a graph.
}
\author{
Ivan Grishagin
}
