% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_graph.R
\name{make_plot_graph}
\alias{make_plot_graph}
\title{Make and Plot BioPAX Graph}
\usage{
make_plot_graph(allnodes, alledges, pw_id = NULL, pw_name = "Pathway",
  to_html = TRUE, to_svg = TRUE)
}
\arguments{
\item{allnodes}{Dataframe with all nodes.}

\item{alledges}{Dataframe with all edges.}

\item{to_html}{Return html file? Has active tooltips.}

\item{to_svg}{Return svg file?}
}
\description{
Using dataframes of nodes and edges make and plot the graph.
}
\author{
Ivan Grishagin
}
