% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathways.R
\name{plot_pathways}
\alias{plot_pathways}
\title{Plot Pathways}
\usage{
plot_pathways(biopax, pw_ids, verbose = FALSE, ...)
}
\arguments{
\item{biopax}{A BioPAX object.}

\item{pw_ids}{IDs of pathways to plot.}

\item{verbose}{Boolean. Show all warnings?}

\item{...}{Optional \code{to_html} and \code{to_svg} parameters passed to \code{make_plot_graph}.}
}
\description{
Takes in a BioPAX object and pathway ids, and  .
}
\author{
Ivan Grishagin
}
